Move class to live with some related classes.
